refactor: change fixture filters stub package
