+++ aligner/src/test/java/org/omegat/gui/align/AlignerTest.java
@@ -42,8 +42,8 @@ import org.omegat.core.Core;
 import org.omegat.core.segmentation.SRX;
 import org.omegat.core.segmentation.Segmenter;
 import org.omegat.filters2.master.FilterMaster;
-import org.omegat.fixtures.filters.ResourceBundleFilter;
-import org.omegat.fixtures.filters.TextFilter;
+import org.omegat.filters.stub.ResourceBundleFilter;
+import org.omegat.filters.stub.TextFilter;
 import org.omegat.gui.align.Aligner.ComparisonMode;
 import org.omegat.util.Language;
