+++ test/fixtures/org/omegat/filters/stub/PoFilter.java
@@ -30,7 +30,7 @@
  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  **************************************************************************/
 
-package org.omegat.fixtures.filters;
+package org.omegat.filters.stub;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
