+++ test/fixtures/org/omegat/filters/stub/XHTMLFilter.java
@@ -24,7 +24,7 @@
  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  **************************************************************************/
 
-package org.omegat.fixtures.filters;
+package org.omegat.filters.stub;
 
 import java.io.File;
 import java.util.List;
