+++ test/fixtures/org/omegat/filters/stub/html2/HTMLFilter2.java
@@ -25,7 +25,7 @@
  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  **************************************************************************/
 
-package org.omegat.fixtures.filters;
+package org.omegat.filters.stub.html2;
 
 import org.htmlparser.Parser;
 import org.htmlparser.util.ParserException;
