+++ test/fixtures/org/omegat/filters/stub/html2/HTMLOptions.java
@@ -26,7 +26,7 @@
  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  **************************************************************************/
 
-package org.omegat.fixtures.filters;
+package org.omegat.filters.stub.html2;
 
 import org.omegat.filters2.AbstractOptions;
