+++ test/fixtures/org/omegat/filters/stub/html2/HTMLReader.java
@@ -24,7 +24,7 @@
  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  **************************************************************************/
 
-package org.omegat.fixtures.filters;
+package org.omegat.filters.stub.html2;
 
 import java.io.BufferedInputStream;
 import java.io.BufferedReader;
