+++ test/src/org/omegat/core/data/ExternalTMFactoryTest.java
@@ -41,8 +41,8 @@ import org.omegat.core.TestCore;
 import org.omegat.core.segmentation.SRX;
 import org.omegat.core.segmentation.Segmenter;
 import org.omegat.filters2.master.FilterMaster;
-import org.omegat.fixtures.filters.MozillaLangFilter;
-import org.omegat.fixtures.filters.PoFilter;
+import org.omegat.filters.stub.MozillaLangFilter;
+import org.omegat.filters.stub.PoFilter;
 import org.omegat.util.Language;
 import org.omegat.util.Preferences;
 import org.omegat.util.TestPreferencesInitializer;
